requirements.txt
aiogram==3.22.0
fastapi>=0.111.0
uvicorn
pydantic>=2.0,<3.0
python-dotenv
